So much stuff already here?
Hi, this is a custom File System, that stores the data in a bin file that can be retrieved automatically, abstracting the whole low-level process of doing so.
Why is there already so much code? This is a File System that I intended to keep integrated with my other project (alongside which this FS was being crafted). However, since we're talking about such a different Scope of either projects, and the vast difference in size, I decided to seperate the two entirely (thus this new repo), unfortunately the story of this project before this repo, will be mixed somewhere in the last project: it's just not worth the effort to shuffle the commits around.